fix: incorrect sync between goroutines
Prior to this commit, goroutines were fetching `NextUnfinishedBlocks`
incorrectly, it was possible for a block to be uploaded more than once,
even though it had no impact on the final upload, the progress of
`totalSent` bytes was overshooting, due to dirty reads.

This commit changes `NextUnfinishedBlocks` to
`ClaimNextUnfinishedBlocks` to atomically mark blocks as "RUNNING" while
the producer fetches it for processing so that a block will not be seen
as unclaimed while a `NextUnfinishedBlocks` happens before the block in
currentt batch has been marked as `RUNNING`.